feat : add admin auth delete and modify publi

+++ api/controllers/publication.js
@@ -1,4 +1,5 @@
 const Publication = require('../models/publication')
+const User = require('../models/user')
 const fs = require('fs') // modules nodeJS permet de créer et gérer des fichiers
 const { runInNewContext } = require('vm')
 
@@ -42,20 +43,41 @@ exports.deletePublication = (req, res, next) => {
     Publication.findOne({ _id: req.params.id }) // On trouve l'objet dans la BDD
         .then((publication) => {
             if (!publication) {
-                res.status(404).json({ error: new Error('Erreur') })
+                return res.status(404).json({ error: new Error('Erreur') })
             }
-            if (publication.userId !== req.auth.userId) {
-                res.status(400).json({ error: new Error('Requète non authorisé!') })
-            }
-            const filename = publication.imageUrl.split('/images/')[1] // Une fois trouvé, on extrait le nom du fichier à supprimer
-            fs.unlink(`images/${filename}`, () => {
-                // On le supprime avec fs.unlink
-                Publication.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(204).json({ message: 'Publication supprimé !' }))
-                    .catch((error) => res.status(400).json({ error }))
-            })
+
+            User.findOne({ _id: req.auth.userId })
+                .then((user) => {
+                    if (!user) {
+                        console.log("L'utilisateur n'existe pas")
+                        return res.status(404).json({ error: 'Utilsateur non trouvé' })
+                    }
+
+                    // Si l'user est le propriétaire de la publication ou s'il est admin, on supprime la publication
+                    if (publication.userId === req.auth.userId || user.admin === true) {
+                        // On delete la publication ...
+                        Publication.deleteOne({ _id: req.params.id })
+                            .then(() => {
+                                // ... si elle est bien effacée, on tente de supprimer l'image
+                                const filename = publication.imageUrl.split('/images/')[1] // Une fois trouvé, on extrait le nom du fichier à supprimer
+
+                                fs.unlink(`images/${filename}`, (err) => {
+                                    if (err) {
+                                        return res.status(500).json({ error: err })
+                                    }
+                                    return res
+                                        .status(200)
+                                        .json({ message: 'Publication supprimée !' })
+                                })
+                            })
+                            .catch((error) => res.status(404).json({ error }))
+                    } else {
+                        return res.status(403).send('unauthorized request')
+                    }
+                })
+                .catch((error) => console.log('raté'))
         })
-        .catch((error) => res.status(400).json({ error }))
+        .catch((error) => res.status(404).json({ error }))
 }
 
 exports.likes = (req, res, next) => {
